+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Browser.pdf
 * Browser.setCache and cache settings
+* DateMock now uses a proxy and supports using it as a function
 * Internal error handling improvements
 * Puppeteer updated to 1.18.1, size and performance improved